- Started creating Databases - Created a little test for searching in Spotify in the Home page - Changed separation between buttons in the CrashHandlerActivity - Added some UI components (SpotifySongCard, VariousChips.kt)

+++ app/src/main/java/com/bobbyesp/spowlo/features/lyrics_downloader/data/local/MediaStoreReceiver.kt
@@ -46,7 +46,7 @@ object MediaStoreReceiver {
                 val title = cursor.getString(titleColumn)
                 val artist = cursor.getString(artistColumn)
                 val album = cursor.getString(albumColumn)
-                val duration = cursor.getInt(durationColumn)
+                val duration = cursor.getDouble(durationColumn)
                 val albumId = cursor.getLong(albumIdColumn)
                 val path = cursor.getString(pathColumn)
